Statistics.LinearRegression

Simple linear regression functions

linearRegression :: Vector v Double => v Double -> v Double -> (Double, Double)

Simple linear regression between 2 samples. Takes two vectors Y={yi} and X={xi} and returns (alpha, beta) such that Y = alpha + beta*X

linearRegressionRSqr :: Vector v Double => v Double -> v Double -> (Double, Double, Double)

Simple linear regression between 2 samples. Takes two vectors Y={yi} and X={xi} and returns (alpha, beta, r*r) such that Y = alpha + beta*X and where r is the Pearson product-moment correlation coefficient

linearRegressionTLS :: Vector v Double => v Double -> v Double -> (Double, Double)

Total Least Squares (TLS) linear regression. Assumes x-axis values (and not just y-axis values) are random variables and that both variables have similar distributions. interface is the same as linearRegression.

Related functions

correl :: Vector v Double => v Double -> v Double -> Double

Pearson's product-moment correlation coefficient

covar :: Vector v Double => v Double -> v Double -> Double

Covariance of two samples

Estimated errors and distribution parameters

linearRegressionMSE :: (Vector v Double, Vector v (Double, Double)) => (Double, Double) -> v Double -> v Double -> Double

The error (or residual) mean square of a sample w.r.t. an estimated regression line. This serves as an estimate for the variance of the sampled data. Accepts the regression parameters (alpha,beta) and the sample vectors X and Y.

linearRegressionDistributions :: (Vector v Double, Vector v (Double, Double)) => (Double, Double) -> v Double -> v Double -> (LinearTransform StudentT, LinearTransform StudentT)

The estimated distributions of the regression parameters (alpha and beta) assuming normal, identical distributions of Y, the sampled data. These can serve to get confidence intervals for the regression parameters. Accepts the regression parameters (alpha,beta) and the sample vectors X and Y. The distributions are StudnetT distributions centered at the estimated (alpha,beta) respectively, with parameter numbers n-2 (where n is the initial sample size) and with standard deviations that are extracted from the sampled data based on its MSE. See chapter 2 of reference [3] for details.

Robust linear regression

robustFit :: (MonadRandom m, Vector v Double) => EstimationParameters -> v Double -> v Double -> m EstimatedRelation

Finding a robust fit linear estimate between two samples. The procedure requires randomization and is based on the procedure described in the reference.

nonRandomRobustFit :: Vector v Double => EstimationParameters -> v Double -> v Double -> EstimatedRelation

A wrapper that executes robustFit using a default random generator (meaning it is only pseudo-random)

robustFitRSqr :: (MonadRandom m, Vector v Double, Vector v (Double, Double)) => EstimationParameters -> v Double -> v Double -> m (EstimatedRelation, Double)

Robust fit yielding also the R-square value of the "clean" dataset.

Related types

data EstimationParameters

The robust fit algorithm used has various parameters that can be specified using the EstimationParameters record.

Constructors

EstimationParameters

Fields

outlierFraction :: !Double

Maximal fraction of outliers expected in the sample (default 0.25)

shortIterationSteps :: !Int

Number of concentration steps to take for initial evaluation of a solution (default 3)

maxSubsetsNum :: !Int

Maximal number of sampled subsets (pairs of points) to use as starting points (default 500)

groupSubsets :: !Int

If the initial sample is large, and thus gets subdivided, this is the number of candidate-estimations to take from each subgroup, on which complete convergence will be executed (default 10)

mediumSetSize :: !Int

Maximal size of sample that can be analyzed without any sub-division (default 600)

largeSetSize :: !Int

Maximal size of sample that does not require two-step sub-division (see reference article) (default 1500)

estimator :: Estimator

Estimator function to use (default linearRegression)

errorFunction :: ErrorFunction

ErrorFunction to use (default linearRegressionError)

type ErrorFunction = EstimatedRelation -> (Double, Double) -> Double

An ErrorFunction is a function that computes the error of a given point from an estimate. This module provides two error functions correspoinding to the two Estimator functions it defines:

Vertical distance squared via linearRegressionError that should be used with linearRegression

Total distance squared vie linearRegressionTLSError that should be used with linearRegressionTLS

type Estimator = Sample -> Sample -> EstimatedRelation

An Estimator is a function that generates an estimated linear regression based on 2 samples. This module provides two estimator functions: linearRegression and linearRegressionTLS

type EstimatedRelation = (Double, Double)

An estimated linear relation between 2 samples is (alpha,beta) such that Y = alpha + beta*X.

Provided values

defaultEstimationParameters :: EstimationParameters

Default set of parameters to use (see reference for details).

linearRegressionError :: ErrorFunction

linearRegression error function is the square of the vertical distance of a point from the line.

linearRegressionTLSError :: ErrorFunction

linearRegressionTLS error function is the square of the total distance of a point from the line.

Helper functions

converge :: Vector v Double => EstimationParameters -> v Double -> v Double -> EstimatedRelation -> EstimatedRelation

Calculate the optimal (local minimum) estimate based on an initial estimate. The local minimum may not be the global (a.k.a. best) estimate but starting from enough different initial estimates should yield the global optimum eventually.

References

Two Dimensional Euclidean Regression (Stein) http://www.dspcsp.com/pubs/euclreg.pdf

Computing LTS Regression For Large Data Sets (Rousseeuw and Driessen) http://agoras.ua.ac.be/abstract/Comlts99.htm

Applied linear statistical models (Kutner et al.)